[#21] Refactor: 유저 데이터 모듈 분리, 디바운싱 코드 수정
- 디바운싱이 제대로 작동하지 않아 우선은 제거
- 유저 데이터 별개 모듈 분리
- 네이밍 수정
